requirements.txt
langchain==0.1.20
langchain-openai==0.1.3 
langchain-groq==0.1.2
langchain-anthropic==0.1.12
openai
python-dotenv